nearest neighbor

+++ source/algorithms/shortest-path/Dijkstra.h
@@ -12,7 +12,7 @@
 class Dijkstra {
     Vertex * s;
     Graph * graph;
-    unordered_map<Vertex *, size_t> dist;
+    unordered_map<Vertex *, long double> dist;
     unordered_map<Vertex *, Vertex *> path;
 public:
     void initialize(Graph * graph, int origin_id);
